+++ apecs/src/Apecs/Experimental/Reactive.hs
@@ -18,53 +18,50 @@ Use e.g. @rget >>= mapLookup True@ to retrieve a list of entities that have a @T
 {-# LANGUAGE ScopedTypeVariables   #-}
 {-# LANGUAGE TypeFamilies          #-}
 
-module Apecs.Experimental.Reactive where
+module Apecs.Experimental.Reactive
+  ( Reacts (..), Reactive, withReactive
+  , EnumMap, enumLookup
+  , OrdMap, ordLookup
+  , IxMap, ixLookup
+  ) where
 
 import           Control.Monad.Reader
-import qualified Data.IntMap.Strict   as M
+import qualified Data.Array.IO        as A
+import qualified Data.IntMap.Strict   as IM
 import qualified Data.IntSet          as S
 import           Data.IORef
+import           Data.Ix
+import qualified Data.Map.Strict      as M
 
-import           Apecs.Core
 import           Apecs.Components
-
--- | Analogous to @Elem@, but for @Reacts@ instances.
---   For a @Reactive r s@ to be valid, @ReactElem r = Elem s@
-type family ReactElem r
+import           Apecs.Core
 
 -- | Class required by @Reactive@.
 --   Given some @r@ and update information about some component, will run a side-effect in monad @m@.
 --   Note that there are also instances for @(,)@.
 class Monad m => Reacts m r where
   rempty :: m r
-  react  :: Entity -> Maybe (ReactElem r) -> Maybe (ReactElem r) -> r -> m ()
-
-type instance ReactElem (a,b) = ReactElem a
-instance (ReactElem a ~ ReactElem b, Reacts m a, Reacts m b) => Reacts m (a, b) where
-  {-# INLINE rempty #-}
-  rempty = liftM2 (,) rempty rempty
-  {-# INLINE react #-}
-  react ety old new (a,b) = react ety old new a >> react ety old new b
+  react  :: Entity -> Maybe (Elem r) -> Maybe (Elem r) -> r -> m ()
 
 -- | Wrapper for reactivity around some store s.
 data Reactive r s = Reactive r s
 
 type instance Elem (Reactive r s) = Elem s
 
--- | Reads @r@ from the game world.
-rget :: forall w m r s.
-  ( Component (ReactElem r)
-  , Has w m (ReactElem r)
-  , Storage (ReactElem r) ~ Reactive r s
-  ) => SystemT w m r
-rget = do
+-- | Performs an action with a reactive state token.
+withReactive :: forall w m r s a.
+             ( Component (Elem r)
+             , Has w m (Elem r)
+             , Storage (Elem r) ~ Reactive r s
+             ) => (r -> m a) -> SystemT w m a
+withReactive f = do
   Reactive r (_ :: s) <- getStore
-  return r
+  lift$ f r
 
 instance (Reacts m r, ExplInit m s) => ExplInit m (Reactive r s) where
   explInit = liftM2 Reactive rempty explInit
 
-instance (Reacts m r, ExplSet m s, ExplGet m s, Elem s ~ ReactElem r)
+instance (Reacts m r, ExplSet m s, ExplGet m s, Elem s ~ Elem r)
   => ExplSet m (Reactive r s) where
   {-# INLINE explSet #-}
   explSet (Reactive r s) ety c = do
@@ -72,7 +69,7 @@ instance (Reacts m r, ExplSet m s, ExplGet m s, Elem s ~ ReactElem r)
     react (Entity ety) old (Just c) r
     explSet s ety c
 
-instance (Reacts m r, ExplDestroy m s, ExplGet m s, Elem s ~ ReactElem r)
+instance (Reacts m r, ExplDestroy m s, ExplGet m s, Elem s ~ Elem r)
   => ExplDestroy m (Reactive r s) where
   {-# INLINE explDestroy #-}
   explDestroy (Reactive r s) ety = do
@@ -92,7 +89,7 @@ instance ExplMembers m s => ExplMembers m (Reactive r s) where
 
 -- | Prints a message to stdout every time a component is updated.
 data Printer c = Printer
-type instance ReactElem (Printer c) = c
+type instance Elem (Printer c) = c
 
 instance (MonadIO m, Show c) => Reacts m (Printer c) where
   {-# INLINE rempty #-}
@@ -107,25 +104,77 @@ instance (MonadIO m, Show c) => Reacts m (Printer c) where
   react _ _ _ _ = return ()
 
 -- | Allows you to look up entities by component value.
---   Use e.g. @rget >>= mapLookup True@ to retrieve a list of entities that have a @True@ component.
-newtype EnumMap c = EnumMap (IORef (M.IntMap S.IntSet))
+--   Use e.g. @withReactive $ mapLookup True@ to retrieve a list of entities that have a @True@ component.
+--   Based on an @IntMap IntSet@ internally.
+newtype EnumMap c = EnumMap (IORef (IM.IntMap S.IntSet))
 
-type instance ReactElem (EnumMap c) = c
+type instance Elem (EnumMap c) = c
 instance (MonadIO m, Enum c) => Reacts m (EnumMap c) where
   {-# INLINE rempty #-}
   rempty = liftIO$ EnumMap <$> newIORef mempty
   {-# INLINE react #-}
   react _ Nothing Nothing _ = return ()
   react (Entity ety) (Just c) Nothing (EnumMap ref) = liftIO$
-    modifyIORef' ref (M.adjust (S.delete ety) (fromEnum c))
+    modifyIORef' ref (IM.adjust (S.delete ety) (fromEnum c))
   react (Entity ety) Nothing (Just c) (EnumMap ref) = liftIO$
-    modifyIORef' ref (M.insertWith mappend (fromEnum c) (S.singleton ety))
+    modifyIORef' ref (IM.insertWith mappend (fromEnum c) (S.singleton ety))
   react (Entity ety) (Just old) (Just new) (EnumMap ref) = liftIO$ do
-    modifyIORef' ref (M.adjust (S.delete ety) (fromEnum old))
-    modifyIORef' ref (M.insertWith mappend (fromEnum new) (S.singleton ety))
+    modifyIORef' ref (IM.adjust (S.delete ety) (fromEnum old))
+    modifyIORef' ref (IM.insertWith mappend (fromEnum new) (S.singleton ety))
 
-{-# INLINE mapLookup #-}
-mapLookup :: Enum c => EnumMap c -> c -> System w [Entity]
-mapLookup (EnumMap ref) c = do
+{-# INLINE enumLookup #-}
+enumLookup :: Enum c => c -> EnumMap c -> System w [Entity]
+enumLookup c = \(EnumMap ref) -> do
   emap <- liftIO $ readIORef ref
-  return $ maybe [] (fmap Entity . S.toList) (M.lookup (fromEnum c) emap)
+  return $ maybe [] (fmap Entity . S.toList) (IM.lookup (fromEnum c) emap)
+
+-- | Allows you to look up entities by component value.
+--   Based on a @Map c IntSet@ internally
+newtype OrdMap c = OrdMap (IORef (M.Map c S.IntSet))
+
+type instance Elem (OrdMap c) = c
+instance (MonadIO m, Ord c) => Reacts m (OrdMap c) where
+  {-# INLINE rempty #-}
+  rempty = liftIO$ OrdMap <$> newIORef mempty
+  {-# INLINE react #-}
+  react _ Nothing Nothing _ = return ()
+  react (Entity ety) (Just c) Nothing (OrdMap ref) = liftIO$
+    modifyIORef' ref (M.adjust (S.delete ety) c)
+  react (Entity ety) Nothing (Just c) (OrdMap ref) = liftIO$
+    modifyIORef' ref (M.insertWith mappend c (S.singleton ety))
+  react (Entity ety) (Just old) (Just new) (OrdMap ref) = liftIO$ do
+    modifyIORef' ref (M.adjust (S.delete ety) old)
+    modifyIORef' ref (M.insertWith mappend new (S.singleton ety))
+
+{-# INLINE ordLookup #-}
+ordLookup :: Ord c => c -> OrdMap c -> System w [Entity]
+ordLookup c = \(OrdMap ref) -> do
+  emap <- liftIO $ readIORef ref
+  return $ maybe [] (fmap Entity . S.toList) (M.lookup c emap)
+
+-- | Allows you to look up entities by component value.
+--   Based on an @IOArray c IntSet@ internally
+newtype IxMap c = IxMap (A.IOArray c S.IntSet)
+
+{-# INLINE modifyArray #-}
+modifyArray :: Ix i => A.IOArray i a -> i -> (a -> a) -> IO ()
+modifyArray ref ix f = A.readArray ref ix >>= A.writeArray ref ix . f
+
+type instance Elem (IxMap c) = c
+instance (MonadIO m, Ix c, Bounded c) => Reacts m (IxMap c) where
+  {-# INLINE rempty #-}
+  rempty = liftIO$ IxMap <$> A.newArray (minBound, maxBound) mempty
+  {-# INLINE react #-}
+  react _ Nothing Nothing _ = return ()
+  react (Entity ety) (Just c) Nothing (IxMap ref) = liftIO$
+    modifyArray ref c (S.delete ety)
+  react (Entity ety) Nothing (Just c) (IxMap ref) = liftIO$
+    modifyArray ref c (S.insert ety)
+  react (Entity ety) (Just old) (Just new) (IxMap ref) = liftIO$ do
+    modifyArray ref old (S.delete ety)
+    modifyArray ref new (S.insert ety)
+
+{-# INLINE ixLookup #-}
+ixLookup :: Ix c => c -> IxMap c -> System w [Entity]
+ixLookup c = \(IxMap ref) -> do
+  liftIO $ fmap Entity . S.toList <$> A.readArray ref c